+++ frontend/src/v5/services/selectorsHooks/currentUserSelectors.hooks.ts
@@ -15,7 +15,7 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import * as CurrentUserSelectors_ from '@/v4/modules/currentUser/currentUser.selectors';
-import { wrapSelectors } from './selectors.helper';
+import * as CurrentUserSelectors from '@/v4/modules/currentUser/currentUser.selectors';
+import { createHooksSelectors } from './selectorsHooks.helper';
 
-export const CurrentUserSelectors = wrapSelectors(CurrentUserSelectors_);
+export const CurrentUserHooksSelectors = createHooksSelectors(CurrentUserSelectors);
